M-3 - 🔧🤖📚✂️ Updating architecture to include a model layer, service layer enhancements, and operational commands; removed old 'core' and 'ai' folders
