--- skinvestigatorai/core/route_generator.py
@@ -1,5 +1,0 @@
-def generate_route_name():
-    import uuid
-    return str(uuid.uuid4()) + '/train'
-
-
