+++ skinvestigatorai/services/custom_image_data_generator.py
