+++ skinvestigatorai/services/data_generator.py
